removed old files, renamed View -> Screen
